feature: team statistics + organize nav routes
