+++ src/data/fbs.json
@@ -41,7 +41,6 @@
 	"Houston Cougars",
 	"Idaho Vandals",
 	"Illinois Fighting Illini",
-	"Illinois State Redbirds",
 	"Indiana Hoosiers",
 	"Iowa Hawkeyes",
 	"Iowa State Cyclones",
